2004-02-27  Al Riddoch  <alriddoch@zepler.org>
	* Fix deps so that we no longer get 2 potentially conflicting
	  versions of wfmath.

	* src/services/gui/widget/Widget.cpp,
	  src/services/gui/widget/Widget.h: Remove some incorrect use
	  of the inline keyword.

	* src/services/metaserver/MetaserverService.cpp,
	  src/services/platform/BitmapRenderer.cpp,
	  src/services/platform/BorderRenderer.cpp: Get rid of
	  some silly warnings.

+++ src/services/metaserver/MetaserverService.cpp
@@ -169,7 +169,7 @@ namespace dime
       << "\n\rServer Type: "<<sInfo.getServer()
       << "\n\rClients: "<<sInfo.getNumClients()
       << " Ping: "<< sInfo.getPing()
-      << " Uptime: "<< sInfo.getUptime()
+      << " Uptime: "<< (int)sInfo.getUptime()
       << ENDM;
 
     return;
